ksmbd-tools: add users who have @group as their primary group
It was reported that Samba supports a use case where @group is treated as a
primary group. Rather than only adding users who have it as a supplementary
group, we now also add users who have the group as their primary group. Since
POSIX does not offer a good way of doing this, we iterate through the system
password database and look for users with the matching GID.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>