ksmbd-tools: prevent recursive @group lookup
A non-standards-compliant login name that begins with an @ sign may
be used e.g. by giving --badname to useradd(8). Users with such names
cannot be used in parameters that take a user list, since the leading
@ sign is reserved for @group, but we still allow such names elsewhere.
If an @group lookup resolves into such a name, treat it now as a user
name rather than another @group. The bug fixed here was introduced in
commit cb106ed ("ksmbd-tools: introduce add_users_map() and stub
add_hosts_map()").

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>