ksmbd-tools: fix more man page formatting quirks
A line with just a comment (`\"') results in a blank line on some
output devices. Fix this by prefixing the comment with an empty request
(`.'). After bold face (`\fB') we go back to previous face (`\fP') or
roman face (`\fR'). The effect here is exactly the same so to avoid
any future confusion, replace all uses of `\fP' with `\fR'. As per
commit e012ed8, add some missing dash escapes.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>